--- base/tests/test_util.py
@@ -1,7 +1,0 @@
-"""
-
-
-"""
-
-import unittest
-from base.util import canonicalize_path
