+++ extract_empirical_results/__init__.py
@@ -1,0 +1,3 @@
+import sys
+if sys.version_info < (3, 6):
+    raise RuntimeError("Requires Python 3.6 or later")
